Moved SelfManagedOnlyException and InternalUserOnlyException to kapua-authorization-api module
