+++ rest-api/core/src/main/java/org/eclipse/kapua/app/api/core/exception/InternalUserOnlyExceptionMapper.java
@@ -13,7 +13,7 @@
 package org.eclipse.kapua.app.api.core.exception;
 
 import org.eclipse.kapua.app.api.core.exception.model.InternalUserOnlyExceptionInfo;
-import org.eclipse.kapua.service.authentication.exception.InternalUserOnlyException;
+import org.eclipse.kapua.service.authorization.exception.InternalUserOnlyException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
