+++ rest-api/core/src/test/java/org/eclipse/kapua/app/api/core/exception/model/SelfManagedOnlyExceptionInfoTest.java
@@ -13,7 +13,7 @@
 package org.eclipse.kapua.app.api.core.exception.model;
 
 import org.eclipse.kapua.qa.markers.junit.JUnitTests;
-import org.eclipse.kapua.service.authentication.exception.SelfManagedOnlyException;
+import org.eclipse.kapua.service.authorization.exception.SelfManagedOnlyException;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
