+++ service/security/authorization/api/src/main/java/org/eclipse/kapua/service/authorization/exception/SelfManagedOnlyException.java
@@ -10,22 +10,24 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.authentication.exception;
+package org.eclipse.kapua.service.authorization.exception;
 
-import org.eclipse.kapua.service.authentication.credential.mfa.MfaOption;
-import org.eclipse.kapua.service.authentication.credential.mfa.MfaOptionService;
-import org.eclipse.kapua.service.authorization.exception.KapuaAuthorizationErrorCodes;
-import org.eclipse.kapua.service.authorization.exception.KapuaAuthorizationException;
+import org.eclipse.kapua.service.user.User;
 
 /**
- * SelfManagedOnlyException is used by the {@link MfaOptionService}, when a user tries to create an {@link MfaOption} for another user.
+ * {@link KapuaAuthorizationException} to {@code throw} when an operation can only be performed on the {@link User} which is currently logged.
  *
- * @since 1.4
+ * @since 1.4.0
  */
 public class SelfManagedOnlyException extends KapuaAuthorizationException {
 
     private static final long serialVersionUID = -3116117876370203330L;
 
+    /**
+     * Constructor.
+     *
+     * @since 1.4.0
+     */
     public SelfManagedOnlyException() {
         super(KapuaAuthorizationErrorCodes.SELF_MANAGED_ONLY);
     }
