+++ service/security/shiro/src/test/java/org/eclipse/kapua/service/authorization/shiro/exception/InternalUserOnlyExceptionTest.java
@@ -13,7 +13,7 @@
 package org.eclipse.kapua.service.authorization.shiro.exception;
 
 import org.eclipse.kapua.qa.markers.junit.JUnitTests;
-import org.eclipse.kapua.service.authentication.exception.InternalUserOnlyException;
+import org.eclipse.kapua.service.authorization.exception.InternalUserOnlyException;
 import org.eclipse.kapua.service.authorization.exception.KapuaAuthorizationErrorCodes;
 import org.junit.Assert;
 import org.junit.Test;
